#429 fix avatar managment
